Moet wel werken natuurlijk

+++ lib/vectorMaths/vectorMaths.h
@@ -32,9 +32,9 @@ typedef struct Vector32 {
  * @brief A floating point three-dimensional vector
  */
 typedef struct Vectorf {
-  int32_t x; /**> x component */
-  int32_t y; /**> y component */
-  int32_t z; /**> z component */
+  float x; /**> x component */
+  float y; /**> y component */
+  float z; /**> z component */
 } Vectorf;
 
 #define clear_vector(a) do {        \
